+++ scripts/integrations/buildABonus.js
@@ -45,9 +45,31 @@ function spellFeature(item) {
 let filters = {
     'spellFeature': spellFeature
 }
+function overlappingEffects(bonuses, object, details, hookType) {
+    if (!bonuses.length) return;
+    let overlappingAuras = new Set();
+    for (let i of bonuses) {
+        if (!i.aura.enabled) continue;
+        if (overlappingAuras.has(i.name)) continue;
+        if (i.bonus === '') continue;
+        let auras = bonuses.filter(b => b.name === i.name && i.bonuses.bonus != '');
+        if (auras.length > 1) overlappingAuras.add(i.name);
+    }
+    if (!overlappingAuras.size) return;
+    let removeUuids = [];
+    for (let i of Array.from(overlappingAuras)) {
+        let evaluatedBonuses = bonuses.filter(b => b.name === i && b.bonuses.bonus != '').map(j => ({'uuid': j.uuid, 'value': new Roll(j.bonuses.bonus, j.getRollData()).evaluate({'async': false, 'maximize': true}).total}));
+        let maxValue = Math.max(...evaluatedBonuses.map(i => i.value));
+        let selectedBonus = evaluatedBonuses.find(i => i.value === maxValue);
+        let removed = evaluatedBonuses.filter(i => i.uuid != selectedBonus.uuid);
+        removeUuids.push(...removed.map(i => i.uuid));
+    }
+    removeUuids.forEach(uuid => bonuses.findSplice(bonus => bonus.uuid === uuid));
+}
 export let buildABonus = {
     'titleBarButton': titleBarButton,
     'daeTitleBarButton': daeTitleBarButton,
     'actorTitleBarButtons': actorTitleBarButton,
-    'filters': filters
+    'filters': filters,
+    'overlappingEffects': overlappingEffects
 }