+++ scripts/macros/classFeatures/cleric/peaceDomain/emboldeningBond.js
@@ -103,7 +103,8 @@ async function checkBonus(effect) {
     }
 }
 async function move(token, changes) {
-    if (game.settings.get('chris-premades', 'LastGM') != game.user.id) return;
+    if (!chris.isLastGM()) return;
+    if (token.parent.id != canvas.scene.id) return;
     if (!changes.x && !changes.y && !changes.elevation) return;
     let effect = chris.findEffect(token.actor, 'Emboldening Bond');
     if (!effect) return;